+++ _docker/db/docker-entrypoint-initdb.d/create_table.sql
@@ -27,7 +27,7 @@ CREATE TABLE couples (
 CREATE TABLE messages (
     id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
     user_id INT UNSIGNED,
-    azure_url VARCHAR(1000) NOT NULL,
+    file_url VARCHAR(1000) NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (user_id) REFERENCES users(id)
 );
