+++ api/v2/azure/storage.py
@@ -2,6 +2,15 @@ from azure.storage.blob import BlobServiceClient
 from sqlalchemy.ext.asyncio import AsyncSession
 
 import v2.models.message as message_model
+from v2.utils.config import get_azure_storage_account, get_azure_sas_token
+
+
+def get_blob_storage_account():
+    azure_storage_account = get_azure_storage_account()
+    account_url = f"https://{azure_storage_account}.blob.core.windows.net"
+    sas_token = get_azure_sas_token()
+    blob_service_client = BlobServiceClient(account_url, credential=sas_token)
+    return blob_service_client
 
 
 async def upload_blob_file(
@@ -18,7 +27,7 @@ async def upload_blob_file(
 
     #messageテーブルにUpdate
     blob_url = blob_client.url
-    message = message_model.Message(user_id=user_id, azure_url=blob_url)
+    message = message_model.Message(user_id=user_id, file_url=blob_url)
     db.add(message)
     await db.commit()
     await db.refresh(message)
