+++ api/v2/cruds/couple.py
@@ -79,22 +79,3 @@ async def update_couple(
 async def delete_couple(db: AsyncSession, original: couple_model.Couple) -> None:
     await db.delete(original)
     await db.commit()
-
-
-async def get_user_id_same_couple(
-    db: AsyncSession, user_id: int
-):
-    #user_idからそのユーザのcouple_idを取得する
-    result: Result = await db.execute(
-        select(user_model.User).where(user_model.User.id == user_id)
-    )
-    user = result.scalar_one_or_none()
-    couple_id = user.couple_id
-
-    #そのcouple_idからもう片方のuser_idを取得する
-    query = select(user_model.User).where(user_model.User.couple_id == couple_id)
-    query = query.where(user_model.User.id != user_id)
-    result = await db.execute(query)
-    user = result.scalar_one_or_none()
-    boddy_id = user.id
-    return boddy_id
